jpa regressions
git-svn-id: svn://svn.caucho.com/resin/trunk@2452 9c94448d-38f1-0310-a231-d98308ff1ebf

+++ modules/resin/src/com/caucho/amber/table/LinkColumns.java
@@ -499,8 +499,8 @@ public class LinkColumns {
         for (Object obj : proxyList) {
           entityType.getHome().getEntityFactory().delete(aConn, obj);
         }
-      }
-      else {
+      } // jpa/0i5e vs. jpa/0h25
+      else if (_sourceTable.getType() == null) {
         CharBuffer cb = new CharBuffer();
 
         cb.append("delete from " + sourceTable +
